# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mall Tenant Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeletedResponse {
  success: boolean;
}

export type TenantStatus = typeof TenantStatus[keyof typeof TenantStatus];


export const TenantStatus = {
  aktif: 'aktif',
  kosong: 'kosong',
  nonaktif: 'nonaktif',
} as const;

export interface Tenant {
  id: number;
  businessName: string;
  ownerName: string;
  email?: string | null;
  phone?: string | null;
  category?: string | null;
  boothNumber?: string | null;
  areaName: string;
  status: TenantStatus;
  createdAt: string;
  updatedAt: string;
}

export type InsertTenantStatus = typeof InsertTenantStatus[keyof typeof InsertTenantStatus];


export const InsertTenantStatus = {
  aktif: 'aktif',
  kosong: 'kosong',
  nonaktif: 'nonaktif',
} as const;

export interface InsertTenant {
  businessName: string;
  ownerName: string;
  email?: string | null;
  phone?: string | null;
  category?: string | null;
  boothNumber?: string | null;
  areaName: string;
  status?: InsertTenantStatus;
}

export type BookingWithTenantPaymentStatus = typeof BookingWithTenantPaymentStatus[keyof typeof BookingWithTenantPaymentStatus];


export const BookingWithTenantPaymentStatus = {
  UNPAID: 'UNPAID',
  PARTIAL: 'PARTIAL',
  PAID: 'PAID',
  OVERDUE: 'OVERDUE',
} as const;

export type BookingWithTenantBookingStatus = typeof BookingWithTenantBookingStatus[keyof typeof BookingWithTenantBookingStatus];


export const BookingWithTenantBookingStatus = {
  aktif: 'aktif',
  selesai: 'selesai',
  pending: 'pending',
  batal: 'batal',
} as const;

export interface BookingWithTenant {
  id: number;
  tenantId: number;
  tenantName: string;
  boothNumber?: string | null;
  areaName: string;
  startDate: string;
  endDate: string;
  totalAmount: number;
  paidAmount: number;
  paymentStatus: BookingWithTenantPaymentStatus;
  bookingStatus: BookingWithTenantBookingStatus;
  dueDate?: string | null;
  periodLabel?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type InsertBookingPaymentStatus = typeof InsertBookingPaymentStatus[keyof typeof InsertBookingPaymentStatus];


export const InsertBookingPaymentStatus = {
  UNPAID: 'UNPAID',
  PARTIAL: 'PARTIAL',
  PAID: 'PAID',
  OVERDUE: 'OVERDUE',
} as const;

export type InsertBookingBookingStatus = typeof InsertBookingBookingStatus[keyof typeof InsertBookingBookingStatus];


export const InsertBookingBookingStatus = {
  aktif: 'aktif',
  selesai: 'selesai',
  pending: 'pending',
  batal: 'batal',
} as const;

export interface InsertBooking {
  tenantId: number;
  startDate: string;
  endDate: string;
  totalAmount: number;
  paidAmount?: number;
  paymentStatus?: InsertBookingPaymentStatus;
  bookingStatus?: InsertBookingBookingStatus;
  dueDate?: string | null;
  periodLabel?: string | null;
}

export type TenantPaymentPaymentMethod = typeof TenantPaymentPaymentMethod[keyof typeof TenantPaymentPaymentMethod];


export const TenantPaymentPaymentMethod = {
  tunai: 'tunai',
  transfer: 'transfer',
  qris: 'qris',
} as const;

export interface TenantPayment {
  id: number;
  bookingId: number;
  amount: number;
  paymentMethod: TenantPaymentPaymentMethod;
  notes?: string | null;
  paidAt: string;
  createdAt: string;
}

export type InsertPaymentPaymentMethod = typeof InsertPaymentPaymentMethod[keyof typeof InsertPaymentPaymentMethod];


export const InsertPaymentPaymentMethod = {
  tunai: 'tunai',
  transfer: 'transfer',
  qris: 'qris',
} as const;

export interface InsertPayment {
  bookingId: number;
  amount: number;
  paymentMethod?: InsertPaymentPaymentMethod;
  notes?: string | null;
}

export interface PosOverview {
  totalActiveTenants: number;
  unpaidCount: number;
  overdueCount: number;
  paidTodayAmount: number;
}

export interface FloorPlanItem {
  id: string;
  tenantId: number;
  bookingId?: number | null;
  businessName: string;
  ownerName: string;
  email?: string | null;
  phone?: string | null;
  category?: string | null;
  boothNumber: string;
  areaName: string;
  startDate?: string | null;
  endDate?: string | null;
  totalAmount: number;
  paidAmount: number;
  remainingAmount: number;
  paymentStatus: string;
  bookingStatus: string;
  dueDate?: string | null;
  periodLabel?: string | null;
}