# lib/api-zod/src/generated/types/bookingWithTenant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mall Tenant Management API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingWithTenantBookingStatus } from './bookingWithTenantBookingStatus';
import type { BookingWithTenantPaymentStatus } from './bookingWithTenantPaymentStatus';

export interface BookingWithTenant {
  id: number;
  tenantId: number;
  tenantName: string;
  boothNumber?: string | null;
  areaName: string;
  startDate: string;
  endDate: string;
  totalAmount: number;
  paidAmount: number;
  paymentStatus: BookingWithTenantPaymentStatus;
  bookingStatus: BookingWithTenantBookingStatus;
  dueDate?: string | null;
  periodLabel?: string | null;
  createdAt: string;
  updatedAt: string;
}