# lib/api-zod/src/generated/types/floorPlanItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mall Tenant Management API
 * OpenAPI spec version: 0.1.0
 */

export interface FloorPlanItem {
  id: string;
  tenantId: number;
  bookingId?: number | null;
  businessName: string;
  ownerName: string;
  email?: string | null;
  phone?: string | null;
  category?: string | null;
  boothNumber: string;
  areaName: string;
  startDate?: string | null;
  endDate?: string | null;
  totalAmount: number;
  paidAmount: number;
  remainingAmount: number;
  paymentStatus: string;
  bookingStatus: string;
  dueDate?: string | null;
  periodLabel?: string | null;
}