# lib/api-zod/src/generated/types/insertBooking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mall Tenant Management API
 * OpenAPI spec version: 0.1.0
 */
import type { InsertBookingBookingStatus } from './insertBookingBookingStatus';
import type { InsertBookingPaymentStatus } from './insertBookingPaymentStatus';

export interface InsertBooking {
  tenantId: number;
  startDate: string;
  endDate: string;
  totalAmount: number;
  paidAmount?: number;
  paymentStatus?: InsertBookingPaymentStatus;
  bookingStatus?: InsertBookingBookingStatus;
  dueDate?: string | null;
  periodLabel?: string | null;
}