# lib/api-zod/src/generated/types/insertPayment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mall Tenant Management API
 * OpenAPI spec version: 0.1.0
 */
import type { InsertPaymentPaymentMethod } from './insertPaymentPaymentMethod';

export interface InsertPayment {
  bookingId: number;
  amount: number;
  paymentMethod?: InsertPaymentPaymentMethod;
  notes?: string | null;
}