# lib/api-zod/src/generated/types/insertTenant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mall Tenant Management API
 * OpenAPI spec version: 0.1.0
 */
import type { InsertTenantStatus } from './insertTenantStatus';

export interface InsertTenant {
  businessName: string;
  ownerName: string;
  email?: string | null;
  phone?: string | null;
  category?: string | null;
  boothNumber?: string | null;
  areaName: string;
  status?: InsertTenantStatus;
}