# lib/api-zod/src/generated/types/posOverview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mall Tenant Management API
 * OpenAPI spec version: 0.1.0
 */

export interface PosOverview {
  totalActiveTenants: number;
  unpaidCount: number;
  overdueCount: number;
  paidTodayAmount: number;
}