# lib/api-zod/src/generated/types/tenantPayment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mall Tenant Management API
 * OpenAPI spec version: 0.1.0
 */
import type { TenantPaymentPaymentMethod } from './tenantPaymentPaymentMethod';

export interface TenantPayment {
  id: number;
  bookingId: number;
  amount: number;
  paymentMethod: TenantPaymentPaymentMethod;
  notes?: string | null;
  paidAt: string;
  createdAt: string;
}